+++ backend-t-shirt/src/main/java/dk/ss/backendtshirt/tshirt/repository/OrderGiftRepository.java
@@ -3,5 +3,5 @@ package dk.ss.backendtshirt.tshirt.repository;
 import dk.ss.backendtshirt.tshirt.model.OrderGift;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderGiftRepository extends JpaRepository<OrderGift, Integer> {
+public interface OrderGiftRepository extends JpaRepository<OrderGift, Long> {
 }
